+++ CHANGELOG.md
@@ -8,6 +8,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## Unreleased
 
+## 7.3.0 - 2024-12-20
+
 ### Added
 
 - `phylum exception` subcommand for managing suppressions